Сделал настройки для каждого профиля
